+++ src/modules/permissions/db/dropped-messages.ts
@@ -1,4 +1,4 @@
-import { getDb } from './connection.js';
+import { getDb } from '../../../db/connection.js';
 
 export interface UnregisteredSender {
   channel_type: string;
